Breaking api and rt store out into separate classes
